move leafinput signature code into leafinputsignature.cpp
