+++ src/BuildLoop.cpp
@@ -211,85 +211,12 @@ static NodeBuildResult::Enum ExecuteNode(BuildQueue* queue, RuntimeNode* node, M
 }
 
 
-static HashDigest ComputeLeafInputSignature(BuildQueue* queue, ThreadState* thread_state, RuntimeNode* node)
-{
-    char path[kMaxPathLength];
-    snprintf(path, sizeof(path), "%s/tmp-%d", queue->m_Config.m_Dag->m_CacheSignatureDirectoryName.Get(), node->m_DagNodeIndex);
-
-    char offlinePath[kMaxPathLength];
-    snprintf(offlinePath, sizeof(offlinePath), "%s/offline-%d", queue->m_Config.m_Dag->m_CacheSignatureDirectoryName.Get(), node->m_DagNodeIndex);
-
-    FILE *sig = fopen(path, "w");
-    FILE *sigoffline = fopen(offlinePath, "r");
-
-    if (sig == NULL)
-        CroakErrno("Failed opening signature ingredients for writing.");
-
-    if (sigoffline == NULL)
-        CroakErrno("Failed opening offline signature ingredients for reading.");
-
-#if TUNDRA_WIN32
-    const char* username = getenv("USERNAME");
-    if (username != nullptr)
-        fprintf(sig, "Signature generated by %s\n", username);
-#else
-    char hostname[1024];
-    if (gethostname(hostname, sizeof(hostname)) == 0)
-        fprintf(sig, "Signature generated by %s\n", hostname);
-#endif
-    time_t rawtime;
-    struct tm *info;
-    time( &rawtime );
-    info = localtime( &rawtime );
-    fprintf(sig, "Current local time and date: %s", asctime(info));
-    fprintf(sig, "Annotation %s\n\n", node->m_DagNode->m_Annotation.Get());
-
-    fprintf(sig, "Cold hash ingredients (anything below this line is hashed and must not change to get a cache hit):\n");
-
-    char            buffer[1024];
-    size_t          n;
-
-    while ((n = fread(buffer, sizeof(char), sizeof(buffer), sigoffline)) > 0)
-    {
-        if (fwrite(buffer, sizeof(char), n, sig) != n)
-            CroakErrno("Failed writing cold signature ingredients.");
-    }
-    fclose(sigoffline);
-
-    fprintf(sig, "Hot hash ingredients:\n");
-    HashDigest res = ComputeLeafInputSignature(
-            queue->m_Config.m_DagNodeIndexToRuntimeNodeIndex_Table,
-            queue->m_Config.m_RuntimeNodes,
-            queue->m_Config.m_Dag,
-            queue->m_Config.m_DagDerived,
-            &queue->m_Config.m_DagRuntimeData,
-            node->m_DagNode,
-            &thread_state->m_LocalHeap,
-            &thread_state->m_ScratchAlloc,
-            thread_state->m_ProfilerThreadId,
-            queue->m_Config.m_StatCache,
-            queue->m_Config.m_DigestCache,
-            queue->m_Config.m_ScanCache,
-            sig);
-    fclose(sig);
-
-    char digestString[kDigestStringSize];
-    DigestToString(digestString, res);
-
-    char new_path[kMaxPathLength];
-    snprintf(new_path, sizeof(path), "%s/%s", queue->m_Config.m_Dag->m_CacheSignatureDirectoryName.Get(), digestString);
-    if (!RenameFile(path, new_path))
-        CroakErrno("Failed moving signature ingredients file.");
-
-    return res;
-}
-
 static bool AttemptToMakeConsistentWithoutNeedingDependenciesBuilt(RuntimeNode* node, BuildQueue* queue, ThreadState* thread_state)
 {
     if (RuntimeNodeHasAttemptedCacheLookup(node))
         return false;
 
-    HashDigest currentLeafInputSignature = ComputeLeafInputSignature(queue, thread_state, node);
+    HashDigest currentLeafInputSignature = CalculateLeafInputSignature(queue, thread_state, node);
     node->m_CurrentLeafInputSignature = currentLeafInputSignature;
 
     if (node->m_BuiltNode)
