+++ src/LeafInputSignature.cpp
@@ -7,6 +7,7 @@
 #include "FileSign.hpp"
 #include "Scanner.hpp"
 #include "Driver.hpp"
+#include <time.h>
 
 static bool FilterOutGeneratedIncludedFiles(void* _userData, const char* includingFile, const char* includedFile)
 {
@@ -14,7 +15,7 @@ static bool FilterOutGeneratedIncludedFiles(void* _userData, const char* includi
     return !IsFileGenerated((DagRuntimeData*)_userData, Djb2HashPath(includedFile), includedFile);
 }
 
-HashDigest ComputeLeafInputSignature(const int32_t* dagNodeIndexToRuntimeNodeIndex_Table, RuntimeNode* runtimeNodesArray, const Frozen::Dag* dag, const Frozen::DagDerived* dagDerived, const DagRuntimeData *dagRuntime, const Frozen::DagNode* dagNode, MemAllocHeap* heap, MemAllocLinear* scratch, int profilerThreadId, StatCache* stat_cache, DigestCache* digest_cache, ScanCache* scan_cache, FILE* ingredient_stream)
+static HashDigest CalculateLeafInputSignature(const int32_t* dagNodeIndexToRuntimeNodeIndex_Table, RuntimeNode* runtimeNodesArray, const Frozen::Dag* dag, const Frozen::DagDerived* dagDerived, const DagRuntimeData *dagRuntime, const Frozen::DagNode* dagNode, MemAllocHeap* heap, MemAllocLinear* scratch, int profilerThreadId, StatCache* stat_cache, DigestCache* digest_cache, ScanCache* scan_cache, FILE* ingredient_stream)
 {
     HashState hashState;
     HashInit(&hashState);
@@ -29,7 +30,7 @@ HashDigest ComputeLeafInputSignature(const int32_t* dagNodeIndexToRuntimeNodeInd
         {
             //this is racy, but it should be okay(tm). It's possible the leaf input signature isn't stored yet, but another thread is already calculating it.
             //in that case we calculate it too, and both threads will just store the same value.
-            childRuntimeNode.m_CurrentLeafInputSignature = ComputeLeafInputSignature(dagNodeIndexToRuntimeNodeIndex_Table, runtimeNodesArray, dag, dagDerived, dagRuntime, &childDagNode, heap, scratch, profilerThreadId, stat_cache, digest_cache, scan_cache, nullptr);
+            childRuntimeNode.m_CurrentLeafInputSignature = CalculateLeafInputSignature(dagNodeIndexToRuntimeNodeIndex_Table, runtimeNodesArray, dag, dagDerived, dagRuntime, &childDagNode, heap, scratch, profilerThreadId, stat_cache, digest_cache, scan_cache, nullptr);
         }
 
         if (ingredient_stream)
@@ -176,6 +177,80 @@ HashDigest CalculateLeafInputHashOffline(const Frozen::Dag* dag, std::function<c
     return hashResult;
 };
 
+
+HashDigest CalculateLeafInputSignature(BuildQueue* queue, ThreadState* thread_state, RuntimeNode* node)
+{
+    char path[kMaxPathLength];
+    snprintf(path, sizeof(path), "%s/tmp-%d", queue->m_Config.m_Dag->m_CacheSignatureDirectoryName.Get(), node->m_DagNodeIndex);
+
+    char offlinePath[kMaxPathLength];
+    snprintf(offlinePath, sizeof(offlinePath), "%s/offline-%d", queue->m_Config.m_Dag->m_CacheSignatureDirectoryName.Get(), node->m_DagNodeIndex);
+
+    FILE *sig = fopen(path, "w");
+    FILE *sigoffline = fopen(offlinePath, "r");
+
+    if (sig == NULL)
+        CroakErrno("Failed opening signature ingredients for writing.");
+
+    if (sigoffline == NULL)
+        CroakErrno("Failed opening offline signature ingredients for reading.");
+
+#if TUNDRA_WIN32
+    const char* username = getenv("USERNAME");
+    if (username != nullptr)
+        fprintf(sig, "Signature generated by %s\n", username);
+#else
+    char hostname[1024];
+    if (gethostname(hostname, sizeof(hostname)) == 0)
+        fprintf(sig, "Signature generated by %s\n", hostname);
+#endif
+    time_t rawtime;
+    struct tm *info;
+    time( &rawtime );
+    info = localtime( &rawtime );
+    fprintf(sig, "Current local time and date: %s", asctime(info));
+    fprintf(sig, "Annotation %s\n\n", node->m_DagNode->m_Annotation.Get());
+
+    fprintf(sig, "Cold hash ingredients (anything below this line is hashed and must not change to get a cache hit):\n");
+
+    char            buffer[1024];
+    size_t          n;
+
+    while ((n = fread(buffer, sizeof(char), sizeof(buffer), sigoffline)) > 0)
+    {
+        if (fwrite(buffer, sizeof(char), n, sig) != n)
+            CroakErrno("Failed writing cold signature ingredients.");
+    }
+    fclose(sigoffline);
+
+    fprintf(sig, "Hot hash ingredients:\n");
+    HashDigest res = CalculateLeafInputSignature(
+            queue->m_Config.m_DagNodeIndexToRuntimeNodeIndex_Table,
+            queue->m_Config.m_RuntimeNodes,
+            queue->m_Config.m_Dag,
+            queue->m_Config.m_DagDerived,
+            &queue->m_Config.m_DagRuntimeData,
+            node->m_DagNode,
+            &thread_state->m_LocalHeap,
+            &thread_state->m_ScratchAlloc,
+            thread_state->m_ProfilerThreadId,
+            queue->m_Config.m_StatCache,
+            queue->m_Config.m_DigestCache,
+            queue->m_Config.m_ScanCache,
+            sig);
+    fclose(sig);
+
+    char digestString[kDigestStringSize];
+    DigestToString(digestString, res);
+
+    char new_path[kMaxPathLength];
+    snprintf(new_path, sizeof(path), "%s/%s", queue->m_Config.m_Dag->m_CacheSignatureDirectoryName.Get(), digestString);
+    if (!RenameFile(path, new_path))
+        CroakErrno("Failed moving signature ingredients file.");
+
+    return res;
+}
+
 void PrintLeafInputSignature(Driver* driver, const char **argv, int argc)
 {
     Buffer<int32_t> requestedNodes;
@@ -207,7 +282,7 @@ void PrintLeafInputSignature(Driver* driver, const char **argv, int argc)
     DagRuntimeData runtimeData;
     DagRuntimeDataInit(&runtimeData, driver->m_DagData, &driver->m_Heap);
 
-    ComputeLeafInputSignature(
+    CalculateLeafInputSignature(
         driver->m_DagNodeIndexToRuntimeNodeIndex_Table.begin(),
         driver->m_RuntimeNodes.begin(),
         driver->m_DagData,
